[Pal] Move Linux elf-x86_64.h to include/arch/x86_64/Linux/elf-arch.h
Also, adapt #includes where needed. Avoid the name elf.h to avoid
clashes. We do not touch the Linux-SGX/elf-x86_64.h file since it is
slightly different.
